backend/requirements.txt
Flask==3.0.3
flask-cors==4.0.1
langchain-groq==0.2.0
langchain-core==0.3.10
supabase==2.4.0
python-dotenv==1.0.1
gspread==6.1.2
google-auth==2.35.0
gunicorn==21.2.0